+++ src/Particle.cpp
@@ -1,19 +1,19 @@
 #include "..\headers\Particle.h"
 
 Particle::Particle() :
-    position(Eigen::Vector3f(0.0f, 0.0f, 0.0f)),
-    velocity(Eigen::Vector3f(0.0f, 0.0f, 0.0f)),
-    acceleration(Eigen::Vector3f(0.0f, 0.0f, 0.0f)),
+    position(Eigen::Vector3d(0.0f, 0.0f, 0.0f)),
+    velocity(Eigen::Vector3d(0.0f, 0.0f, 0.0f)),
+    acceleration(Eigen::Vector3d(0.0f, 0.0f, 0.0f)),
 
     mass(SPACING * SPACING * SPACING * REST_DENSITY),
     density(REST_DENSITY),
     pressure(0.0f),
     isFluid(true),
     ID(0),
-    predictedVelocity(Eigen::Vector3f(0.0f, 0.0f, 0.0f)),
+    predictedVelocity(Eigen::Vector3d(0.0f, 0.0f, 0.0f)),
     sourceTerm(0.0f),
     diagonal(0.0f),
-    pressureAcceleration(Eigen::Vector3f(0.0f, 0.0f, 0.0f)) {}
+    pressureAcceleration(Eigen::Vector3d(0.0f, 0.0f, 0.0f)) {}
 
 Eigen::Vector3i Particle::getCellNumber() {
 	int x = std::floor(position.x() / CELL_SIZE);
@@ -23,26 +23,26 @@ Eigen::Vector3i Particle::getCellNumber() {
 }
 
 Particle2D::Particle2D() :
-	position(Eigen::Vector2f(0.0f, 0.0f)),
-	velocity(Eigen::Vector2f(0.0f, 0.0f)),
-	acceleration(Eigen::Vector2f(0.0f, 0.0f)),
-	normal(Eigen::Vector2f(0.0f, 0.0f)),
+	position(Eigen::Vector2d(0.0, 0.0)),
+	velocity(Eigen::Vector2d(0.0, 0.0)),
+	acceleration(Eigen::Vector2d(0.0, 0.0)),
+	normal(Eigen::Vector2d(0.0, 0.0)),
 
 	mass(SPACING * SPACING * REST_DENSITY),
 	density(REST_DENSITY),
-	pressure(0.0f),
+	pressure(0.0),
 	isFluid(true),
 	isSurface(false),
 	ID(0),
-	predictedVelocity(Eigen::Vector2f(0.0f, 0.0f)),
-	sourceTerm(0.0f),
-	diagonal(0.0f),
-	pressureAcceleration(Eigen::Vector2f(0.0f, 0.0f)),
-	dii(Eigen::Vector2f(0.0f, 0.0f)),
+	predictedVelocity(Eigen::Vector2d(0.0, 0.0)),
+	sourceTerm(0.0),
+	diagonal(0.0),
+	pressureAcceleration(Eigen::Vector2d(0.0, 0.0)),
+	dii(Eigen::Vector2d(0.0f, 0.0f)),
 	aii(0.0f),
-	predictedDensity(0.0f),
-	predictedPressure(0.0f),
-	ci(Eigen::Vector2f(0.0f, 0.0f)) {}
+	predictedDensity(0.0),
+	predictedPressure(0.0),
+	ci(Eigen::Vector2d(0.0f, 0.0f)) {}
 
 Eigen::Vector2i Particle2D::getCellNumber() {
 	int x = std::floor(position.x() / CELL_SIZE);
